switch some uint32_t into size_t

+++ includes/tacopie/error.hpp
@@ -9,7 +9,7 @@ namespace tacopie {
 class error : public std::runtime_error {
 public:
   //! ctor & dtor
-  error(const std::string& what, const std::string& file, std::uint32_t line);
+  error(const std::string& what, const std::string& file, std::size_t line);
   ~error(void) = default;
 
   //! copy ctor & assignment operator
@@ -19,12 +19,12 @@ public:
 public:
   //! get location of the error
   const std::string& get_file(void) const;
-  std::uint32_t get_line(void) const;
+  std::size_t get_line(void) const;
 
 private:
   //! location of the error
   std::string m_file;
-  std::uint32_t m_line;
+  std::size_t m_line;
 };
 
 } //! tacopie
